--- apps/meteor/app/threads/client/index.ts
@@ -1,2 +1,0 @@
-import './messageAction/useReplyInThreadMessageAction';
-import './threads.css';
