+++ apps/meteor/client/views/room/contextualBar/Threads/components/ThreadMessageList.tsx
@@ -21,6 +21,7 @@ import { useMessageListNavigation } from '../../../hooks/useMessageListNavigatio
 import { useLegacyThreadMessageJump } from '../hooks/useLegacyThreadMessageJump';
 import { useLegacyThreadMessageListScrolling } from '../hooks/useLegacyThreadMessageListScrolling';
 import { useLegacyThreadMessages } from '../hooks/useLegacyThreadMessages';
+import './threads.css';
 
 const isMessageSequential = (current: IMessage, previous: IMessage | undefined, groupingRange: number): boolean => {
 	if (!previous) {
